requirements.txt
APScheduler==3.6.3
asgiref==3.5.0
cachetools==4.2.2
certifi==2021.10.8
chardet==3.0.4
click==8.0.4
Django==4.0.3
googletrans==3.1.0a0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
joblib==1.1.0
nltk==3.7
python-dotenv==0.19.2
python-telegram-bot==13.11
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.3.15
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
textblob==0.17.1
tornado==6.1
tqdm==4.63.0
tzdata==2021.5
tzlocal==4.1
